don't print stupid-sized date
Wordpress would print a huge date before a post if the date was different
from the (newer) post ahead of it.  When I first converted from wp to poc,
it was easier to *always* output the date.  If one steps back and thinks
about it, printing the date in such huge font is not just needless, but also
distracting.  (It also wastes a level of headings making \subsection and
\subsubsection headings look small.)  Drop the huge date heading, and
instead print the date right next to the time.