daemon: drop unneeded privileges on startup
If someone manages to do bad things, at least they'll be limited to
reading/writing files, and accessing the network.  Of course, it'd be great
if this could be locked down further - existing data can be only read, new
comments can be only written, and network access is only allowed on
incomming connections.  Alas, I don't think that's possible on Illumos.