template: finish support for simplified if-markup
Instead of allowing template language functions to take arbitrary number of
args, we allow zero, one, or two.  Currently support functions are:

	ifgt(x,y)
	iflt(x,y)
	ifeq(x,y)
	else()
	endif()

The if-functions take two arguments.  If the arguments is an integer, the
value is used as is.  If the argument is a string, it's used to look up a
variable of that name.  If the variable doesn't exist, the value zero is
used.  If the variable exists, but doesn't contain an integer value an
assertion is triggered.  Finally, if it exists and the value is an integer,
the value is used as is.  Once the x and y values are parsed and extracted,
simple integer comparison is performed.