str: store sufficiently small strings inline in struct str
This is one of the older tricks in the book - if the string pointed to by
the structure is small enough, just stash it in the structure itself.  If
that is the common case, then the savings should be significant.  Since
we're using struct str on per token basis, this saves a significant number
of generic malloc()-based allocations.