replace all the yaml code with lisp code
It turns out yaml is nice but way too complicated.  libyaml is a huge pain
to use because it doesn't provide a nice way to "just convert this buffer
into a structure I can query".  Instead it provides a way to process a
stream of tokens.  This is unfortunate and it requires quite a bit of code.

We can use lisp-like syntax for the meta data files and stop relying on
libyaml and all the code to wrap it.

Signed-off-by: Josef 'Jeff' Sipek <jeffpc@josefsipek.net>