sidebar: remove archive listing from the sidebar
The archive list was getting way too long and fundamentally it is a useless
form of navigation.  All it conveys is that during a certain month, there
has been a post published.  It doesn't say how many posts or what the posts
were about.

Note: This commit does *not* remove support for the ?m=YYYYMM URLs.  They
still work as before, they are just never linked to anymore.

Signed-off-by: Josef 'Jeff' Sipek <jeffpc@josefsipek.net>