post index: introduce the post index
In order to service requests properly and promptly, we need to index all the
posts in several different ways.  To make things easy on ourselves, this
commit introduces the post index.  Whenever we allocate a post, we insert it
into all the indices via a single simple function call.  The index code
takes care of adding references to the post from all the subindices.

Signed-off-by: Josef 'Jeff' Sipek <jeffpc@josefsipek.net>