Added tag v3.1 for changeset e94fb945e386
Again, so much has changed.

The major change is the migration from CGI to SCGI.  This caused us to make
two daemons: blahgd and mathd.  mathd is responsible for generating math
files.  blahgd is responsible for everything else.  The two communicate
using a door call.  This brings up a shift in focus.  For the longest time,
the code tried to be portable between POSIX systems.  That's no longer the
case.  Now it only targets Illumos-based systems.  Some of the Illumos
features
used are:

  - doors
  - libavl
  - libnvpair
  - libcmdutils (for list_t & friends)
  - libumem
  - process privileges

In the process, of converting to a daemon model, we had to fix tons of
memory leaks, and provide a mechanism for blahgd to serve static files.
There are also SMF manifests for the two daemons.  Finally, the code should
now be both 32- and 64-bit clean.

Visitor visible changes include change in font, the favicon, availability of
comment feeds, and a math captcha to avoid comment spam (see README for
details).