Added tag v3.2 for changeset d35b9749eb26
Conditional templates

Yep, after years of living without conditional markup in the template
language, I finally bit the bullet and implemented simple conditional logic.
It doesn't support much, but it supports enough to get rid of the two
biggest annoyances: the "previous page" and "next page" links going off to
infinity, and posts with no comments failing to validate because of an empty
<ol> tag.

Other than than, a couple of memory leaks were addressed, the str and req
structures are now allocated using umem caches, and finally the requests
keep track of assorted latencies.