Added tag v3.3 for changeset 96baabdc65fd
MIT license

After way too much thought, I decided that the MIT license is the way to go.
This is the first version that has any license on it.  Quite exciting!

Other than that, the one big (technical) change is the logging of every
request.  I expect this to get useless/tiresome, but until then...disk is
cheap.