Added tag v4.1 for changeset 1f7ff43e4321
There are two major changes since 4.0.  First of all, we're now using
libjeffpc to handle a lot of the generic data handling.  The blahgd repo
contained a lot of useful wrappers and helpers, and so those got moved into
libjeffpc so that other projects can benefit.  Second, we removed support
for fmt 1 and fmt 2 posts.  This leaves support for fmt 3 only.