utilities & tests: no longer need an explicit libjeffpc initialization
Utilities and tests don't need the explicit initialization.  There were two
reasons blahg code did it:

(1) memory allocation purposes - making sure that the libjeffpc slab
    allocator was initialized
(2) setting up a log function - to direct errors to syslog

Recently, libjeffpc changed so that it uses shared object constructors,
therefore #1 is no longer a requirement.  And as far as #2 is concerned, the
utilities and tests don't want to spam syslog anyway.  Previously, they
accomplished this by setting BLAHG_DISABLE_SYSLOG=1.  Since not logging is
the default for libjeffpc, we can accomplish the same thing by not giving it
the custom ops vector and relying on the defaults.

Signed-off-by: Josef 'Jeff' Sipek <jeffpc@josefsipek.net>