post index: don't list unlisted posts
We can set the "listed" metadata field to false to avoid having the post
show up in all of the lists of posts - index, archive, tags, and categories.
(This includes the feeds since they are really just different format
rendering of the same content.)

This is not meant as a security feature.  Anyone who knows the post id can
just request it specifically.

Signed-off-by: Josef 'Jeff' Sipek <jeffpc@josefsipek.net>