req: move rendering latency info into a HTTP header
Instead of appending a HTML comment on certain outputs about how long the
rendering of the page took, unconditionally inject a X-blahg-* header with
the timing information.

This fixes a regression introduced by the switch away from custom stats
keeping that resulted in HTML pages having to wait for a timeout in the HTTP
server.

Signed-off-by: Josef 'Jeff' Sipek <jeffpc@josefsipek.net>