post fmt3: don't wrap verbatim environment in <p></p>
The whole point of the verbatim environment is to allow inserting arbitrary
HTML into the output.  Therefore, it doesn't make sense to wrap the result
in a paragraph tag.  Furthermore, this change avoids HTML validator errors
caused by illegal nesting of tags.  For example, a hand-crafted table in a
verbatim environment produces:

	<p><table>...</table></p>

This is an error because <table> cannot exist inside a paragraph.  With this
change, the outer <p> isn't added making a valid HTML document.

Signed-off-by: Josef 'Jeff' Sipek <jeffpc@josefsipek.net>